rteval: code cleanup and fixes
Changes:

* Closed /dev/null after use on each function opening it.

* Replaced os.kill(PID, SIGTERM) by subprocess.terminate() as the processes
  have been started using subprocess.Popen(). Also added subprocess.wait()
  after each terminate() call.
  Then I noticed the subprocess python module in RHEL5 doesn't have the
  methods terminate() and kill(), so I reverted that piece of the change
  but kept the missing wait() calls.

* Imported the missing SIGKILL symbol in hackbench.py

* Removed the code snippet below from hackbench.py:runload():

         count = 30
         while count > 0 and p.poll() == None:
             time.sleep(1.0)
             count -= 1
         if p.poll() == None:
             os.kill(p.pid, SIGKILL)

+++ rteval/kcompile.py
@@ -97,6 +97,7 @@ class Kcompile(load.Load):
             return
         self.debug("ready to run")
         self.ready = True
+        os.close(null)
 
     def runload(self):
         null = os.open("/dev/null", os.O_RDWR)
@@ -118,7 +119,10 @@ class Kcompile(load.Load):
                 p = subprocess.Popen(self.args,
                                      stdin=null,stdout=null,stderr=null)
         self.debug("stopping")
-        os.kill(p.pid, SIGTERM)
+        if p.poll() == None:
+            os.kill(p.pid, SIGTERM)
+        p.wait()
+        os.close(null)
 
     def genxml(self, x):
         x.taggedvalue('command_line', ' '.join(self.args), {'name':'kcompile'})
